+++ features/step_definitions/track_step.rb
@@ -6,6 +6,9 @@ And(/watch folder is created if it does not exists/) do
   watch_dir = Miyuki.config['watchDir']
   FileUtils.rm_r(watch_dir) if Dir.exists?(watch_dir)
 
+  track_file = Miyuki.config['trackFile']
+  FileUtils.rm(track_file) if File.exists?(track_file)
+
   expect { Miyuki.track! }.to change { Dir.exists?(watch_dir) }.from(false).to(true)
 end